https://github.com/Necrobot-Private/NecroBot/tree/c401f8e8912ef115685a614ddb9d55448da806d5

+++ PoGo.NecroBot.Logic/Tasks/MSniperServiceTask.cs
@@ -1,6 +1,7 @@
 ﻿using GeoCoordinatePortable;
 using Microsoft.AspNet.SignalR.Client;
 using Newtonsoft.Json;
+using PoGo.NecroBot.Logic.Common;
 using PoGo.NecroBot.Logic.Event;
 using PoGo.NecroBot.Logic.Exceptions;
 using PoGo.NecroBot.Logic.Logging;
@@ -31,6 +32,7 @@ namespace PoGo.NecroBot.Logic.Tasks
         public static List<string> VisitedEncounterIds = new List<string>();
         private static List<MSniperInfo2> autoSnipePokemons = new List<MSniperInfo2>();
         private static List<MSniperInfo2> manualSnipePokemons = new List<MSniperInfo2>();
+        private static List<MSniperInfo2> pokedexSnipePokemons = new List<MSniperInfo2>();
         private static bool inProgress = false;
         private static DateTime OutOffBallBlock = DateTime.MinValue;
         public static bool isConnected = false;
@@ -197,6 +199,7 @@ namespace PoGo.NecroBot.Logic.Tasks
 
         public class MSniperInfo2
         {
+            public DateTime AddedTime { get;  set; }
             public ulong EncounterId { get; set; }
             public double Iv { get; set; }
             public double Latitude { get; set; }
@@ -240,6 +243,59 @@ namespace PoGo.NecroBot.Logic.Tasks
             });
         }
 
+        public static void UnblockSnipe(bool spinned = true)
+        {
+            snipeFailedCount = 0;
+            waitNextPokestop = spinned;
+        }
+        static DateTime lastPrintMessageTime = DateTime.Now;
+
+        private static bool CheckSnipeConditions(ISession session)
+        {
+            if (waitNextPokestop) return false;
+            if (session.LoggedTime > DateTime.Now.AddMinutes(1)) return false; //only snipe after login 1 min.
+
+            if (snipeFailedCount >= 3) return false;
+            if (session.Stats.CatchThresholdExceeds(session)) return false;
+
+            if (inProgress || OutOffBallBlock > DateTime.Now)
+                return false;
+
+            if (!session.LogicSettings.UseSnipeLimit) return true;
+
+            if (lastPrintMessageTime.AddMinutes(1) > DateTime.Now)
+            {
+                session.EventDispatcher.Send(new SnipeEvent
+                {
+                    Message = session.Translation.GetTranslation(TranslationString.SniperCount, session.Stats.SnipeCount)
+                });
+
+            }
+            if (session.Stats.LastSnipeTime.AddMilliseconds(session.LogicSettings.MinDelayBetweenSnipes) > DateTime.Now) return false;
+
+            if (session.Stats.SnipeCount < session.LogicSettings.SnipeCountLimit)
+                return true;
+
+            if ((DateTime.Now - session.Stats.LastSnipeTime).TotalSeconds > session.LogicSettings.SnipeRestSeconds)
+            {
+                session.Stats.SnipeCount = 0;
+            }
+            else
+            {
+                if (lastPrintMessageTime.AddMinutes(1) > DateTime.Now)
+                {
+
+                    lastPrintMessageTime = DateTime.Now;
+                    session.EventDispatcher.Send(new SnipeEvent
+                    {
+                        Message = session.Translation.GetTranslation(TranslationString.SnipeExceeds)
+                    });
+                }
+                return false;
+            }
+            return true;
+        }
+
         public static async Task<bool> CatchFromService(ISession session, CancellationToken cancellationToken, MSniperInfo2 encounterId)
         {
             cancellationToken.ThrowIfCancellationRequested();
@@ -263,6 +319,10 @@ namespace PoGo.NecroBot.Logic.Tasks
             {
                 throw ex;
             }
+            catch(Exception ex)
+            {
+                return false;
+            }
             finally
             {
                 await LocationUtils.UpdatePlayerLocationWithAltitude(session,
@@ -271,6 +331,7 @@ namespace PoGo.NecroBot.Logic.Tasks
 
             if (encounter.Status == EncounterResponse.Types.Status.PokemonInventoryFull)
             {
+                Logger.Write("Pokemon bag full, snipe cancel");
                 return false;
             }
             PokemonData encounteredPokemon;
@@ -328,16 +389,37 @@ namespace PoGo.NecroBot.Logic.Tasks
 
         #endregion MSniper Location Feeder
 
-        public static void AddSnipeItem(ISession session, MSniperInfo2 item, bool byPassValidation = false)
+        public static async Task AddSnipeItem(ISession session, MSniperInfo2 item, bool byPassValidation = false)
         {
+            item.AddedTime = DateTime.Now;
+            //just keep pokemon in last 2 min
+            autoSnipePokemons.RemoveAll(x => x.AddedTime.AddMinutes(2) < DateTime.Now);
+            pokedexSnipePokemons.RemoveAll(x => x.AddedTime.AddMinutes(2) < DateTime.Now);
+
             if (OutOffBallBlock > DateTime.Now ||
+                
                 autoSnipePokemons.Exists(x => x.EncounterId == item.EncounterId && item.EncounterId > 0) ||
-                (item.EncounterId > 0 && session.Cache[item.EncounterId.ToString()] != null)) return;
+                (item.EncounterId > 0 && session.Cache[item.EncounterId.ToString()] != null) ||
+                (session.LogicSettings.SnipeIgnoreUnknownIv && item.Iv == 0 && !byPassValidation) ) return;
 
             item.Iv = Math.Round(item.Iv, 2);
+            if (session.LogicSettings.SnipePokemonNotInPokedex)
+            {
+                var pokedex = await session.Inventory.GetPokeDexItems();
+
+                if (!pokedex.Exists(x => x.InventoryItemData?.PokedexEntry?.PokemonId == (PokemonId)item.PokemonId) && 
+                    !pokedexSnipePokemons.Exists(p=>p.PokemonId == item.PokemonId) &&
+                    (!session.LogicSettings.AutosnipeVerifiedOnly || (session.LogicSettings.AutosnipeVerifiedOnly && item.EncounterId >0)))
+                {
+                    session.EventDispatcher.Send(new WarnEvent() { Message = session.Translation.GetTranslation(TranslationString.SnipePokemonNotInPokedex, session.Translation.GetPokemonTranslation((PokemonId)item.PokemonId)) });
+                    pokedexSnipePokemons.Add(item);//Add as hight priority snipe entry
+
+                }
+            }
             SnipeFilter filter = new SnipeFilter()
             {
-                SnipeIV = session.LogicSettings.MinIVForAutoSnipe
+                SnipeIV = session.LogicSettings.MinIVForAutoSnipe   ,
+                VerifiedOnly = session.LogicSettings.AutosnipeVerifiedOnly
             };
 
             var pokemonId = (PokemonId)item.PokemonId;
@@ -350,7 +432,8 @@ namespace PoGo.NecroBot.Logic.Tasks
             if (byPassValidation)
             {
                 manualSnipePokemons.Add(item);
-                Logger.Write($"(MANUAL SNIPER) Pokemon added |  {item.PokemonId} [{item.Latitude},{item.Longitude}] IV {item.Iv}%");
+                
+                Logger.Write($"(MANUAL SNIPER) Pokemon added |  {(PokemonId)item.PokemonId} [{item.Latitude},{item.Longitude}] IV {item.Iv}%");
                 return;
             }
 
@@ -386,19 +469,19 @@ namespace PoGo.NecroBot.Logic.Tasks
             }
         }
 
-        public static async Task CatchWithSnipe(ISession session, MSniperInfo2 encounterId, CancellationToken cancellationToken)
+        public static async Task<bool> CatchWithSnipe(ISession session, MSniperInfo2 encounterId, CancellationToken cancellationToken)
         {
             cancellationToken.ThrowIfCancellationRequested();
-            await
+            return await
                   SnipePokemonTask.Snipe(session, new List<PokemonId>() { (PokemonId)encounterId.PokemonId }, encounterId.Latitude, encounterId.Longitude, cancellationToken);
         }
 
+        static int snipeFailedCount = 0;
+        static bool waitNextPokestop = true;
         public static async Task Execute(ISession session, CancellationToken cancellationToken)
         {
-            if (session.Stats.CatchThresholdExceeds(session)) return;
+            if (!CheckSnipeConditions(session)) return;
 
-            if (inProgress || OutOffBallBlock > DateTime.Now)
-            return;
             inProgress = true;
 
             var pth = Path.Combine(Directory.GetCurrentDirectory(), "SnipeMS.json");
@@ -431,21 +514,49 @@ namespace PoGo.NecroBot.Logic.Tasks
                     File.Delete(pth);
                     if (mSniperLocation2 == null) mSniperLocation2 = new List<MSniperInfo2>();
                 }
+                if(pokedexSnipePokemons.Count >0)
+                {
+                    mSniperLocation2.Add(pokedexSnipePokemons.OrderByDescending(x=>x.PokemonId).FirstOrDefault());
+                    pokedexSnipePokemons.Clear();
+
+                }
                 if (manualSnipePokemons.Count > 0)
                 {
+                    
                     mSniperLocation2.AddRange(manualSnipePokemons);
                     manualSnipePokemons.Clear();
                 }
                 else {
-                    autoSnipePokemons.Reverse();
-                    mSniperLocation2.AddRange(autoSnipePokemons.Take(10));
-                    autoSnipePokemons.Clear();
+
+                    autoSnipePokemons.RemoveAll(x => x.AddedTime.AddMinutes(2) < DateTime.Now);
+                    autoSnipePokemons.OrderByDescending(x => x.PokemonId).ThenByDescending(x => x.AddedTime);
+                    var batch = autoSnipePokemons.Take(5);
+                    //mSniperLocation2.AddRange(autoSnipePokemons.Take(10));
+                    //autoSnipePokemons.Clear();
+                    if (batch!= null && batch.Count()>0)
+                    {
+                        mSniperLocation2.AddRange(batch);
+                        autoSnipePokemons.RemoveAll(x=> batch.Contains(x));
+                    }
                 }
                 foreach (var location in mSniperLocation2)
                 {
                     if (session.Stats.CatchThresholdExceeds(session)) break;
 
+                    if (pokedexSnipePokemons.Count > 0) break;
+
                     if (location.EncounterId > 0 && session.Cache[location.EncounterId.ToString()] != null) continue;
+
+                    if (!await SnipePokemonTask.CheckPokeballsToSnipe(session.LogicSettings.MinPokeballsWhileSnipe + 1, session, cancellationToken))
+                    {
+                        session.EventDispatcher.Send(new WarnEvent()
+                        {
+                            Message = session.Translation.GetTranslation(Common.TranslationString.AutoSnipeDisabled)
+                        });
+
+                        OutOffBallBlock = DateTime.Now.AddMinutes(5);
+                        break;
+                    }
 
                     //If bot already catch the same pokemon, and very close this location. 
                     string uniqueCacheKey = $"{session.Settings.PtcUsername}{session.Settings.GoogleUsername}{Math.Round(location.Latitude, 6)}{location.PokemonId}{Math.Round(location.Longitude, 6)}";
@@ -462,19 +573,21 @@ namespace PoGo.NecroBot.Logic.Tasks
                         Source = "MSniperService",
                         Iv = location.Iv
                     });
-                    if (location.EncounterId != 0)
+                    var result = location.EncounterId != 0 ? await CatchFromService(session, cancellationToken, location) : await CatchWithSnipe(session, location, cancellationToken);
+
+                    if (result)
                     {
-                        if (!await CatchFromService(session, cancellationToken, location))
-                        {
-                            //inventory full, out of ball break snipe
-                            break;
-                        }
+                        snipeFailedCount = 0;
                     }
                     else
                     {
-                        await CatchWithSnipe(session, location, cancellationToken);
+                        snipeFailedCount ++;
+                        if (snipeFailedCount >= 3) break; //maybe softban, stop snipe wait until verify it not been 
                     }
                     await Task.Delay(1000, cancellationToken);
+                    session.Stats.LastSnipeTime = DateTime.Now;
+                    session.Stats.SnipeCount++;
+                    waitNextPokestop = true;
                 }
             }
             catch (ActiveSwitchByRuleException ex)
@@ -487,6 +600,8 @@ namespace PoGo.NecroBot.Logic.Tasks
             }
             catch (Exception ex)
             {
+                if (ex.InnerException != null && ex.InnerException is CaptchaException) throw ex.InnerException;
+
                 File.Delete(pth);
                 var ee = new ErrorEvent { Message = ex.Message };
                 if (ex.InnerException != null) ee.Message = ex.InnerException.Message;
